+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **BREAKING CHANGES**:
 
 - **Global**. Added static `.$use()`. Call it before using element: `WUPTextControl.$use()` (_self-registration by import_ works now and will be removed in the future)
+- [PopupElement](src/popup/popupElement.ts). Removed `$isOpen` in favor of `$isShown`
 
 **Fixes**: